refactor: centralize shared dependency facades
