+++ src/cli/features/fs-artifacts.js
@@ -1,0 +1,1 @@
+export * from "../shared/fs-artifacts.js";
