+++ src/cli/features/manual/file-helpers.js
@@ -1,4 +1,4 @@
 export {
     writeFileArtifact as writeManualFile,
     writeJsonArtifact as writeManualJsonArtifact
-} from "../../shared/fs-artifacts.js";
+} from "../fs-artifacts.js";
