+++ src/cli/shared/module-support.js
@@ -1,0 +1,2 @@
+export { getErrorMessage, isErrorWithCode } from "../../shared/utils/error.js";
+export { assertNonEmptyString } from "../../shared/utils/string.js";
