+++ src/semantic/src/dependencies.js
@@ -1,0 +1,1 @@
+export * from "../../shared/index.js";
