+++ src/semantic/src/project-index/abort-guard.js
@@ -1,4 +1,4 @@
-import { createAbortGuard } from "../../../shared/index.js";
+import { createAbortGuard } from "../dependencies.js";
 
 export const PROJECT_ROOT_DISCOVERY_ABORT_MESSAGE =
     "Project root discovery was aborted.";
