+++ src/semantic/src/project-index/concurrency.js
@@ -1,4 +1,4 @@
-import { createEnvConfiguredValueWithFallback } from "../../../shared/index.js";
+import { createEnvConfiguredValueWithFallback } from "../dependencies.js";
 
 const PROJECT_INDEX_GML_CONCURRENCY_ENV_VAR = "GML_PROJECT_INDEX_CONCURRENCY";
 const PROJECT_INDEX_GML_CONCURRENCY_BASELINE = 4;
