+++ src/semantic/src/project-index/coordinator.js
@@ -1,6 +1,6 @@
 import path from "node:path";
 
-import { throwIfAborted } from "../../../shared/index.js";
+import { throwIfAborted } from "../dependencies.js";
 
 function normalizeEnsureReadyDescriptor(descriptor) {
     const projectRoot = descriptor?.projectRoot;
