+++ src/semantic/src/project-index/path-info.js
@@ -3,7 +3,7 @@ import path from "node:path";
 import {
     isNonEmptyString,
     resolveContainedRelativePath
-} from "../../../shared/index.js";
+} from "../dependencies.js";
 
 /**
  * Resolve high-level metadata about how {@link filePath} relates to
