+++ src/semantic/tests/project-index-metrics.test.js
@@ -5,7 +5,7 @@ import path from "node:path";
 import test from "node:test";
 
 import { buildProjectIndex } from "../src/project-index/index.js";
-import { createMetricsTracker } from "../../shared/index.js";
+import { createMetricsTracker } from "../src/dependencies.js";
 
 async function writeProjectFile(rootDir, relativePath, contents) {
     const absolutePath = path.join(rootDir, relativePath);
